Move existing formatted interface into separate file
